+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - New function `Generate Airports`: Extracts the airports from a `Flightplans.txt` file, and uses an airports master file (containing every possible airport entry) to create this flightplan's `Airports.txt` file.
   - The file path to the master file must be provided in the `fs-ai-tools.generateAirports.masterAirportsFilePath` setting.
 - Show Aircraft: Better formatting in popup dialog, using title and separate lines
+- Clean Flightplan: fixed flight number padding if followed by an airport starting with "F" or "R"
 - Increased minimum VSCode version to v1.64.0
 
 ### 1.2.2