+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+### 1.3.1
+- Rebase AC#s: new settings to define the minimum number of empty lines needed to start a new number group. See `fs-ai-tools.rebaseAircraftNumbers.groupSeparationAircraftTxt` and `fs-ai-tools.rebaseAircraftNumbers.groupSeparationFlightplansTxt`.
+
 ### 1.3.0
 - New function `Generate Airports`: Extracts the airports from a `Flightplans.txt` file, and uses an airports master file (containing every possible airport entry) to create this flightplan's `Airports.txt` file.
   - The file path to the master file must be provided in the `fs-ai-tools.generateAirports.masterAirportsFilePath` setting.